1 - Added ctf-invasion 2 - Added Cheney's new ctf code(commented out until sounds made) 3 - Fixed bug with rscript and colormap.

+++ source/ref_gl/r_script.h
@@ -55,8 +55,8 @@ typedef struct
 // colormap
 typedef struct
 {
-	qboolean enabled;
-	int	red, green, blue;	// colors - duh!
+	qboolean	enabled;
+	float		red, green, blue;	// colors - duh!
 } rs_colormap_t;
 
 typedef struct
